requirements.txt
# --- Parameter-Efficient Fine-Tuning (LoRA) ---
peft>=0.11.0
transformers>=4.40.0

# --- OpenAI CLIP repo ---
git+https://github.com/openai/CLIP.git

# --- Utilities ---
tqdm
pandas>=2.0.0
packaging>=23.1

# --- Image processing ---
Pillow>=9.0.0

# --- Metrics ---
scikit-learn>=1.3.0

# --- Visualization ---
matplotlib>=3.7.0
seaborn>=0.12.0

# --- Logging ---
tensorboard>=2.14.0
